# docs/swagger.yaml
definitions:
  base.Blame:
    enum:
    - User
    - Postgres
    - Server
    type: string
    x-enum-varnames:
    - BlameUser
    - BlamePostgres
    - BlameServer
  base.ResponseFailure:
    properties:
      blame:
        allOf:
        - $ref: '#/definitions/base.Blame'
        example: Guilty System
      message:
        example: error occurred
        type: string
      status:
        example: Error
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
    type: object
  base.ResponseOK:
    properties:
      status:
        example: OK
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
    type: object
  base.ResponseOKWithID:
    properties:
      ID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
      status:
        example: OK
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
    type: object
  enum.Level:
    enum:
    - None
    - Beginner
    - Elementary
    - Intermediate
    type: string
    x-enum-varnames:
    - None
    - Beginner
    - Elementary
    - Intermediate
  enum.Role:
    enum:
    - Admin
    - Guest
    type: string
    x-enum-varnames:
    - Admin
    - Guest
  model.GetUserResponse:
    properties:
      status:
        example: OK
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
      user:
        $ref: '#/definitions/model.UserObject'
    type: object
  model.GetUsersResponse:
    properties:
      status:
        example: OK
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
      users:
        items:
          $ref: '#/definitions/model.UserObject'
        type: array
    type: object
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.LoginResponse:
    properties:
      refreshToken:
        type: string
      status:
        example: OK
        type: string
      token:
        type: string
      trackingID:
        example: 12345678-1234-1234-1234-000000000000
        type: string
    type: object
  model.RecreateJWTRequest:
    properties:
      refreshToken:
        type: string
    type: object
  model.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.UpdateUserAllFieldRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      level:
        $ref: '#/definitions/enum.Level'
      password:
        type: string
    type: object
  model.UpdateUserAuthorizationFieldsRequest:
    properties:
      email:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
  model.UserObject:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      level:
        $ref: '#/definitions/enum.Level'
      name:
        type: string
      role:
        $ref: '#/definitions/enum.Role'
      updatedAt:
        type: string
    type: object
  model.UsersByIdListRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /user/:user-id/update:
    post:
      consumes:
      - application/json
      description: Update User All Fields
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserAllFieldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base.ResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      security:
      - ApiKeyAuth: []
      summary: Update User All Fields
      tags:
      - User
  /user/authorizationFields/update:
    post:
      consumes:
      - application/json
      description: Update User Authorization Fields
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserAuthorizationFieldsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base.ResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      security:
      - ApiKeyAuth: []
      summary: Update User Authorization Fields
      tags:
      - User
  /user/get:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUsersResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: User authorisation
      parameters:
      - description: User request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      summary: User authorisation
      tags:
      - Auth
  /user/logout:
    post:
      consumes:
      - application/json
      description: Unauthorized users
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: User request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RecreateJWTRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base.ResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      security:
      - ApiKeyAuth: []
      summary: Unauthorized users
      tags:
      - Auth
  /user/refresh:
    post:
      consumes:
      - application/json
      description: Re-create refresh token
      parameters:
      - description: User request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RecreateJWTRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      summary: Re-create refresh token
      tags:
      - Auth
  /user/register:
    post:
      consumes:
      - application/json
      description: User registration
      parameters:
      - description: User request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base.ResponseOKWithID'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      summary: User registration
      tags:
      - Auth
  /user/retrieve:
    get:
      consumes:
      - application/json
      description: Retrieve data of an authorised user
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      security:
      - ApiKeyAuth: []
      summary: Retrieve data of an authorised user
      tags:
      - User
  /usersByIdList:
    post:
      consumes:
      - application/json
      description: Retrieve user information by id list
      parameters:
      - description: User data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UsersByIdListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUsersResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/base.ResponseFailure'
        "500":
          description: Internal error (server fault)
          schema:
            $ref: '#/definitions/base.ResponseFailure'
      summary: Retrieve user information by id list
      tags:
      - Users
swagger: "2.0"